Completed a version of internal kriging (still needs work, but it cleans the data for now, and it can be used to compare with other approaches).
Stage a) of the MCMC according to Gelfand (2005) seems to be complete.

Begun work on the clustering sections with the new understanding of clustering.

Began to review other approaches to spatial dirichlet process (hence the presence of external work in the code). All externally sourced code should be and will be referenced.